Move TradeBasicDetails components to separated file
